project/requirements.txt
biopython
matplotlib
seaborn
pandas
numpy
scipy 